refactor: Extract common code for deployment scripts
