+++ src/lambdas/zip.py
@@ -1,24 +1,10 @@
 import zipfile
 import os
 from typing import List, Tuple
+from common_deploy import *
 
 
 COMMON_FNAME = "./common.py"
-
-
-def get_lambdas() -> List[Tuple[str, str]]:
-    """
-    Get filenames of all lambdas in this directory.
-    By convention, the module must be a .py and must begin with "lambda_".
-    """
-
-    fnames = []
-    for file in os.listdir("."):
-        if file.startswith("lambda_") and file.endswith(".py"):
-            fullpath = os.path.join(".", file)
-            just_name = file.split(".")[0]
-            fnames.append((fullpath, just_name))
-    return fnames
 
 
 def zip_file(fname: Tuple[str, str]):
@@ -34,7 +20,7 @@ def zip_file(fname: Tuple[str, str]):
 
 
 def main():
-    fnames = get_lambdas()
+    fnames = get_lambda_module_files()
 
     for fname_pair in fnames:
         zip_file(fname_pair)
